+++ archivosjson/1.json
@@ -1,0 +1,1 @@
+{"1": [-0.0846373587846756, 0.08049627393484116, 0.1031511127948761, -0.04891175031661987, -0.028876909986138344, 0.052416954189538956, -0.017503734678030014, -0.08978176862001419, 0.09900315850973129, -0.13922828435897827, 0.33310720324516296, 0.012952238321304321, -0.26569807529449463, -0.16180911660194397, 0.06744210422039032, 0.09710761159658432, -0.21350640058517456, -0.11821141839027405, -0.11161189526319504, -0.0018363073468208313, 0.05248377099633217, -0.05687122792005539, 0.004070992115885019, -0.0069241998717188835, -0.1206565573811531, -0.3927384912967682, -0.024104606360197067, -0.05945466458797455, 0.042465630918741226, -0.12208966165781021, -0.0010093403980135918, -0.030200982466340065, -0.15447619557380676, -0.03346824645996094, -0.06097518652677536, 0.01398206315934658, -0.0387558788061142, -0.13342151045799255, 0.22951844334602356, -0.04837323725223541, -0.23441515862941742, 0.03657402843236923, 0.051879774779081345, 0.21228715777397156, 0.15407578647136688, 0.04979170113801956, 0.025889761745929718, -0.08807462453842163, 0.17096593976020813, -0.27096423506736755, 0.023607119917869568, 0.13648244738578796, 0.13114142417907715, 0.01004346739500761, 0.07526492327451706, -0.18039435148239136, 0.014624198898673058, 0.1409837007522583, -0.22777971625328064, 0.027590030804276466, -0.00889650546014309, -0.09035682678222656, 0.010119806975126266, -0.0775199830532074, 0.22015398740768433, 0.11336846649646759, -0.12966583669185638, -0.1457502543926239, 0.1283389925956726, -0.09264569729566574, -0.09099392592906952, 0.06729524582624435, -0.09573608636856079, -0.20471352338790894, -0.3171917796134949, 0.04539657011628151, 0.35301709175109863, 0.15504449605941772, -0.20538510382175446, 0.022863369435071945, -0.12633943557739258, -0.0523211732506752, 0.07169011980295181, 0.1352352350950241, -0.07302588224411011, -0.023191101849079132, -0.10677798092365265, -0.035036712884902954, 0.16585499048233032, 0.03953541815280914, -0.03814651072025299, 0.23756375908851624, 0.03800847381353378, -0.00043439026921987534, 0.023178433999419212, 0.007305022794753313, -0.1680951565504074, 0.029898259788751602, -0.13262434303760529, -0.05583919584751129, 0.0626697763800621, -0.058284107595682144, 0.05101686716079712, 0.13484810292720795, -0.23679983615875244, 0.16330744326114655, -0.025459416210651398, 0.00195235013961792, 0.12651427090168, -0.01369263045489788, -0.031182317063212395, -0.04031557962298393, 0.06810834258794785, -0.3460842967033386, 0.13603529334068298, 0.20194140076637268, 0.11747240275144577, 0.16538411378860474, 0.1402471512556076, 0.04688306897878647, -0.006910966709256172, 0.014570750296115875, -0.24396789073944092, -0.06503130495548248, 0.05368884652853012, -0.01808372512459755, 0.09662751853466034, 0.023383885622024536]}
