+++ archivosjson/2.json
@@ -1,0 +1,1 @@
+{"2": [-0.131480410695076, 0.1493949443101883, 0.05456921458244324, -0.09057066589593887, -0.08131761848926544, -0.005273470655083656, -0.05266006290912628, -0.08337009698152542, 0.16827277839183807, -0.16045133769512177, 0.1924174427986145, -0.08789288252592087, -0.26158976554870605, -0.051608700305223465, -0.022817790508270264, 0.1668122112751007, -0.23075443506240845, -0.10362858325242996, -0.021949492394924164, -0.07294802367687225, 0.11024734377861023, -0.012963073328137398, -0.015929054468870163, 0.06406703591346741, -0.15003731846809387, -0.32164859771728516, -0.08730387687683105, -0.060954101383686066, 0.003964085131883621, -0.09136112034320831, -0.0672135055065155, -0.07297638803720474, -0.1727449595928192, -0.052987705916166306, -0.051031991839408875, 0.05226338654756546, -0.05194583535194397, -0.10132152587175369, 0.16143397986888885, 0.011401860043406487, -0.16902700066566467, -0.04456496238708496, -0.008419159799814224, 0.2140868604183197, 0.16080355644226074, 0.07266661524772644, 0.04291011393070221, -0.09853048622608185, 0.13444095849990845, -0.19942262768745422, 0.05728403106331825, 0.13361376523971558, 0.08265902101993561, 0.06530804932117462, 0.006182494107633829, -0.18546855449676514, 0.003814193420112133, 0.161770761013031, -0.18566837906837463, 0.07531504333019257, 0.0669993907213211, -0.0906805694103241, 0.054134272038936615, -0.0030819051899015903, 0.23471276462078094, 0.10280488431453705, -0.12019980698823929, -0.15135042369365692, 0.18947243690490723, -0.13501937687397003, -0.05601269379258156, 0.014983479864895344, -0.08776972442865372, -0.22779929637908936, -0.3170037567615509, 0.0560331828892231, 0.3764694333076477, 0.17534546554088593, -0.12675073742866516, -0.025754263624548912, -0.07721849530935287, -0.04024583846330643, 0.10758812725543976, 0.142137348651886, -0.1010199785232544, 0.030110515654087067, -0.10027095675468445, 0.04642467200756073, 0.2140386998653412, -0.012824482284486294, -0.09360677748918533, 0.23759566247463226, 0.002724088728427887, 0.05423024296760559, 0.024476731196045876, 0.03975250944495201, -0.1408253014087677, 0.04945969954133034, -0.07646416127681732, 0.009930686093866825, 0.0659755989909172, -0.11256653070449829, -0.005871788132935762, 0.14741501212120056, -0.15233586728572845, 0.15549753606319427, -0.03449662774801254, -0.014901487156748772, -0.10591458529233932, 0.0052568367682397366, -0.16265054047107697, -0.014459432102739811, 0.15673284232616425, -0.2152550369501114, 0.16576258838176727, 0.22790756821632385, 0.10595575720071793, 0.18622228503227234, 0.1352347731590271, 0.04010903835296631, 0.06167783588171005, -0.05006706714630127, -0.17853495478630066, -0.046796154230833054, 0.08178967982530594, -0.01068797241896391, 0.16590191423892975, 0.07086724787950516]}
