+++ archivosjson/3.json
@@ -1,0 +1,1 @@
+{"3": [-0.09073023498058319, 0.09333066642284393, 0.039376623928546906, 0.05568613484501839, -0.09985321015119553, -0.09907220304012299, -0.01887836679816246, -0.20870819687843323, 0.06999576836824417, -0.1590452492237091, 0.17631466686725616, -0.05852515995502472, -0.24242368340492249, 0.0027882724534720182, -0.0184044037014246, 0.11617513746023178, -0.20852616429328918, -0.20481854677200317, -0.12236038595438004, -0.11490969359874725, 0.01267959550023079, 0.10075146704912186, 0.016266804188489914, 0.035353001207113266, -0.1522115170955658, -0.2888932228088379, -0.05376477912068367, -0.16175445914268494, 0.10850825905799866, -0.08384040743112564, 0.013537653721868992, 0.012099385261535645, -0.1336624175310135, 0.009858520701527596, 0.04438798129558563, 0.006323017179965973, -0.039706259965896606, -0.05772349238395691, 0.24500760436058044, 0.019862040877342224, -0.20137245953083038, -0.026737459003925323, -0.006025921553373337, 0.2954311668872833, 0.188766747713089, -0.03733772411942482, 0.05142252892255783, -0.17734411358833313, 0.08577953279018402, -0.25288254022598267, 0.096694216132164, 0.1267925351858139, 0.12824596464633942, 0.10180202126502991, 0.05460411310195923, -0.1516454517841339, 0.03956340253353119, 0.026557665318250656, -0.23548106849193573, 0.09565135836601257, 0.15162840485572815, -0.08671487867832184, -0.020388348028063774, 0.032252371311187744, 0.18493276834487915, 0.0363084115087986, -0.07147924602031708, -0.18058963119983673, 0.1338980793952942, -0.30527403950691223, -0.16439832746982574, 0.08031485974788666, -0.02502177096903324, -0.15808752179145813, -0.35578471422195435, -0.04138917475938797, 0.3843555748462677, 0.15829786658287048, -0.20348137617111206, 0.048374243080616, -0.03813556209206581, -0.07174914330244064, 0.13248255848884583, 0.1188332587480545, -0.09280470013618469, -0.00991879589855671, -0.10990762710571289, 0.04812796786427498, 0.22226496040821075, -0.049195293337106705, -0.10346511006355286, 0.2104002982378006, 0.0358881875872612, -0.038665276020765305, 0.12178752571344376, 0.07784387469291687, -0.11201995611190796, -0.055523909628391266, -0.062077753245830536, -0.04949456453323364, 0.05699273198843002, -0.01333686150610447, -0.01621711067855358, 0.11851568520069122, -0.1506335735321045, 0.266030877828598, -0.08241647481918335, -0.0745270699262619, -0.029448531568050385, -0.06540921330451965, -0.09575329720973969, -0.024237558245658875, 0.17391836643218994, -0.23501816391944885, 0.162897989153862, 0.17735296487808228, 0.056400999426841736, 0.14828556776046753, 0.007689329795539379, 0.0007631741464138031, -0.03981047868728638, 0.016268428415060043, -0.20687875151634216, -0.10417348146438599, 0.04527595639228821, -0.031108733266592026, 0.047799691557884216, 0.03467005491256714]}
